+++ src/slow/gradient/gradient.py
@@ -5,14 +5,17 @@
 # Author: Y.Takahashi, Hokkaido University
 # Date; 2022/04/07
 
+import logging
 import numpy as np
 from slow.orbital.orbital import orbital
+
+logger = logging.getLogger(__name__)
 
 class gradient(orbital):
 
 
   def __init__(self):
-    print("Calling class: gradient")
+    logger.info('Calling class: gradient')
 
     # Gradient variables
 
@@ -22,7 +25,7 @@ class gradient(orbital):
   @orbital.time_measurement_decorated
   def initialize_gradient(self, config, dimension_dict, geom_dict):
 
-    print('Setting initial gradient variables')
+    logger.info('Setting initial gradient variables')
 
     num_spatial  = 3
     num_primitiv = dimension_dict['num_primitive']
@@ -31,14 +34,14 @@ class gradient(orbital):
     var_gradient = np.zeros(num_spatial*num_primitiv*num_cell).reshape(num_spatial ,num_primitiv, num_cell)
 
 
-    print('--Checking gradient method')
+    logger.info('--Checking gradient method')
     kind_gradient = str( config['gradient_setting']['kind_gradient'] )
     flag_kind_gradient = False
     for n in range(0, len(self.avail_gradient_scheme) ):
       if kind_gradient == self.avail_gradient_scheme[n]:
         flag_kind_gradient = True
     if not flag_kind_gradient:
-      print('Error, gradient scheme is not implemented. Please check the contronl file')
+      logger.info('Error, gradient scheme is not implemented. Please check the contronl file')
       exit()
 
 
@@ -47,14 +50,14 @@ class gradient(orbital):
     var_limiter[:,:] = 1.0
     var_neig_maxmin  = np.zeros(2*num_primitiv*num_cell).reshape(2,num_primitiv, num_cell)
 
-    print('--Checking limiter model')
+    logger.info('--Checking limiter model')
     kind_limiter = str( config['gradient_setting']['kind_limiter'] )
     flag_kind_limiter = False
     for n in range(0, len(self.avail_slope_limiter) ):
       if kind_limiter == self.avail_slope_limiter[n]:
         flag_kind_limiter = True
     if not flag_kind_limiter:
-      print('Error, slope limiter is not implemented. Please check the contronl file')
+      logger.info('Error, slope limiter is not implemented. Please check the contronl file')
       exit()
 
     return var_gradient, var_limiter, var_neig_maxmin